PriceCalculator helper in de AppointmentService gezet
